Refactor test paths
