+++ src/test/java/org/wso2/carbon/module/swiftiso20022/mt/parsers/MTParserTests.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.module.swiftiso20022.mt.parser;
+package org.wso2.carbon.module.swiftiso20022.mt.parsers;
 
 import org.mockito.internal.matchers.apachecommons.ReflectionEquals;
 import org.powermock.core.classloader.annotations.PowerMockIgnore;
@@ -26,8 +26,6 @@ import org.wso2.carbon.module.swiftiso20022.exceptions.MTMessageParsingException
 import org.wso2.carbon.module.swiftiso20022.mt.models.blocks.ApplicationHeaderBlock;
 import org.wso2.carbon.module.swiftiso20022.mt.models.blocks.BasicHeaderBlock;
 import org.wso2.carbon.module.swiftiso20022.mt.models.messages.MT940Message;
-import org.wso2.carbon.module.swiftiso20022.mt.parsers.MT940Parser;
-import org.wso2.carbon.module.swiftiso20022.mt.parsers.MTParser;
 import org.wso2.carbon.module.swiftiso20022.utils.MTParserTestConstants;
 import org.wso2.carbon.module.swiftiso20022.utils.MTParserUtils;
